[backend/강소라][W4][Task3,4] Rental-RentalHisoty, Reservation Rest API 작성 (#282)
* test

* test2

* Revert "test2"

This reverts commit 1b518d1335e61e1462f2ca505322cabd771cfa76.

* [backend/sorakang][#95] doc: 도서관 프로젝트 디렉터리 추가 , Readme 추가

* [backend/sorakang][#95] feat: eslint , 개발 환경 setting

* [backend/sorakang][#95] feat: db dir setting

* [backend/sorakang][#95] feat: db connect and layer 구조 설정

* [backend/sorakang][#95] feat: set environment globally

* [backend/sorakang][#95] feat: git ignore  env.* add

* [backend/sorakang][#95] fix : clear git cache

* test

* [backend/sorakang][#166] fest: model 추가 , db table 생성

* [backend/sorakang][#166] fest: modle 관계 수정, 보완

* [backend/sorakang][#166] fest: seeders 추가 (book , user), db 스키마 수정

* [backend/sorakang][#166] fest: getBookDataApi 설명 추가

* [backend/sorakang][#166] fest: env.example 추가

* [backend/sorakang][#166] fix: READM 추가, dotenv 정의 및 호출 형식 변경

* [backend/sorakang][#166] model 변수명 카멜 -> 파스칼 변경, model 일부 관계 수정, comment 상세하게 수정

* [backend/sorakang][#166] fix: timestamap옵셥 true 제거 : Date type인 경우 default로 적용됨

* [backend/sorakang][#166] fix:
- 리뷰반영
- Model 구조 변경 - 사용자 권한 부여 table 전체적인 수정 (auth 삭제)  , loginInfo Table 생성

* [backend/sorakang][#166] feat : env.example , readme 수정 , models/index.js option 추가

* [backend/sorakang][#211] feat: user rest api pr 추가

* [backend/sorakang][#211] feat: router , controller 추가

* [backend/sorakang][#211] feat: User service logic added

* [backend/sorakang][#95] doc: Readme 에 api문서 추가

* [backend/sorakang][#166] fix: review 반영

* [backend/sorakang][#166] fix: book model 수정

* [backend/sorakang][#166] fix: model eslint 예외 주석 삭제

* [backend/sorakang][#168] feat: user seed 추가 , getuser, getSingleIUser 수정

* [backend/sorakang][#211] feat : User Create 로직 추가 , pw encryption middleware 추가

* [backend/sorakang][#211] feat : User update , delete 로직 추가

* [backend/sorakang][#211] feat : User 로직 형식 통일

* [backend/sorakang][#211] feat : user login logout  added , accessToken

* [backend/sorakang][#211] feat : Repository 구조 변경

* [backend/sorakang][#211] feat : auth middleware added

* [backend/sorakang][#211] feat : refresh token 로직 추가 -- 진행중

* [backend/sorakang][#211] feat : logout 로직 수정

* [backend/sorakang][#211] fix : 간단 수정

* [backend/sorakang][#213] feat: rental rest api pr 추가

* [backend/sorakang][#213] feat: book rental Read , Create 로직 추가

* [backend/sorakang][#211 fix : error handling 변경 (임시)

* [backend/sorakang][#213] feat: rental CRUD 진행 (수정필)

* [backend/sorakang][#213] feat: rent return (진행중)

* [backend/sorakang][#213] feat: reservation 기본 CRUD

* [backend/sorakang][#211] feat : Reservation , Rental 보완 수정

* [backend/강소라][W3][Task6] Rest API - Book CRUD  (#176)

* [backend/sorakang][#168] test

* [backend/sorakang][#168] feat: book Router 추가

* [backend/sorakang][#168] feat: book controller , Service 구조 구축

* [backend/sorakang][#168] feat: book Read 로직 추가

* [backend/sorakang][#168] feat: book 조회 - pagination 추가 , Repository layer 제거 --> 본 프로젝트 규모가 크지 않으므로 MVC pattern 으로 진행

* [backend/sorakang][#168] feat: book 조회 - repository 추가 , 로직 수정 => 추후 transaction처리 고려

* [backend/sorakang][#168] feat: singleBook 조회 기본 로직 추가

* [backend/sorakang][#168] feat: book create 로직 추가 -> bulk create는 추후 구축 예정

* [backend/sorakang][#168] feat: book simple update, delete 로직 추가

* [backend/sorakang][#168] feat: book bulk update 로직 구성 완료

* [backend/sorakang][#168] issue: 콘퍼런스 수정사항 반영 - 조금 더 수정 필요

* [backend/sorakang][#168] feat: book 조회 로직 기능 추가 : pubDate 기준 pagenation 구현

* [backend/sorakang][#168] feat: BOOK CRUD 정리 , book seed 파일 수정

* [backend/sorakang][#168] fix: error code 수정

* [backend/sorakang][#168] fix : Review 반영 - error code , getBookApi 수정

* [backend/sorakang][#168] fix : Review 반영 - updateBook PUT ---> PATCH 변경


* practice commit

* [backend/sorakang][#166] fix : review 반영 - db/models의 require Model 제거

* [backend/sorakang][#211 feat : refresh token 추가 , auth 일부 수정

* [backend/sorakang][#211 fix : delete file

* [backend/sorakang][#211 feat : 중복 로그인 방지 authMiddleware에 추가

* [backend/sorakang][#211] fix : bookRepository.js 수정 : model 컬럼 명 변경 반영, Router에 authMiddleware 추가

* [backend/sorakang][#211] fix : userService error처리 수정

* [backend/sorakang][#211] fix : error 해결

* [backend/sorakang][#213] fix : 주석 해제

* [backend/sorakang][#231] fix : seed 수정 , 추가 , book service 오타 수정

* [backend/sorakang][#213] feat : book & Rental, Reservation 로직 일부 수정

* [backend/sorakang][#213] fix: 리뷰 반영 : reservationController,  retry 함수 수정

* [backend/sorakang][#213] fix: 리뷰 반영 : getReservation -> getAllReservation 수정

* [backend/sorakang][#213] fix: 리뷰 반영 : 변수명 변경 resolveObj => reservationInfo

* [backend/sorakang][#213] fix: 리뷰 반영 : model 필드명 수정 pubDate => publicationDate

* [backend/sorakang][#213] fix: 리뷰 반영 : book delete api 분리 , 수정된 변수명과 의종성 있는 부분 반영

* [backend/sorakang][#213] fix: 리뷰 반영 : bookRental 로직 수정 , 주석 추가

* [backend/sorakang][#213] fix: 리뷰 반영 : bookReservation 로직 수정 , 주석 추가

* [backend/sorakang][#213] fix: 리뷰 반영 : rentalRouter 수정 : authMiddleware 중복성 관련 . /all endpoint 수정

* [backend/sorakang][#213] fix: 리뷰 반영 : reservation , rental

* [backend/sorakang][#211] fix : Review 반영 aToken, rToken -> accessToken, refreshToken 수정, 오타 수정

* [backend/sorakang][#166] fix : review 반영 -  모델 명 camelcase 반영

+++ backend/sorakang/koa-books/src/controller/bookContoller.js
@@ -1,6 +1,6 @@
 const { bookService } = require('../services');
 
-const getBook = async (ctx) => {
+const getAllBook = async (ctx) => {
   try {
     // 입력 값에 대한 Validation이 필요함. 타입에 대한 validation , err -> 해당 주차에서 진행
     const {
@@ -11,7 +11,7 @@ const getBook = async (ctx) => {
     const dateCursor = new Date(Number(cursorArray[0]));
     const bookId = Number(cursorArray[1]);
 
-    ctx.body = await bookService.getBook(Number(limit), dateCursor, bookId, search);
+    ctx.body = await bookService.getAllBook(Number(limit), dateCursor, bookId, search);
   } catch (err) {
     throw new Error(err); // 임시 error handling
   }
@@ -34,12 +34,10 @@ const getSingleBook = async (ctx) => {
 };
 
 const createBook = async (ctx) => {
-  // 입력 값에 대한 Validation이 필요함. 타입에 대한 validation , err -> 해당 주차에서 진행
-
   try {
     const { bookList } = ctx.request.body;
 
-    if (!bookList.length) { /* error handling 필요 */ }
+    if (!bookList?.length === 0) { /* error handling 필요 */ }
 
     const body = await bookService.createBook(bookList);
 
@@ -55,7 +53,7 @@ const updateBook = async (ctx) => {
     const { bookId } = ctx.params;
 
     if (!bookId || !bookInfo) {
-      throw new Error(404, 'No Found');
+      throw new Error(404, 'Invalid request body or query');
     }
 
     const { state, message } = await bookService.updateBook(bookId, bookInfo);
@@ -69,15 +67,10 @@ const updateBook = async (ctx) => {
 
 const deleteBook = async (ctx) => {
   try {
-    const { bookId } = ctx.params;
+    const idList = ctx.request.body.bookIdList;
 
-    const idList = [];
-    if (!ctx.request.body) {
-      idList.push(...ctx.request.body.bookIdList);
-    } else if (bookId) {
-      idList.push(bookId);
-    } else {
-      throw new Error(400, 'Invalid request body or query'); // 임시 error handling
+    if (!idList) {
+      throw new Error(404, 'Invalid request body ');
     }
 
     const result = await bookService.deleteBook(idList);
@@ -91,6 +84,26 @@ const deleteBook = async (ctx) => {
   }
 };
 
+const deleteSingleBook = async (ctx) => {
+  try {
+    const { bookId } = ctx.params;
+
+    if (!bookId) {
+      throw new Error(404, 'Invalid request params ');
+    }
+
+    const result = await bookService.deleteSingleBook(bookId);
+
+    if (result === 0) {
+      throw new Error(404, 'Not Found');
+    }
+
+    ctx.body = { message: 'successfully deleted' };
+  } catch (err) {
+    throw new Error(err.message);/// 임시 error handling
+  }
+};
+
 module.exports = {
-  getBook, getSingleBook, createBook, updateBook, deleteBook,
+  getAllBook, getSingleBook, createBook, updateBook, deleteBook, deleteSingleBook,
 };
